define nft account service to work with different data sources
