+++ src/endpoints/esdt/entities/esdt.data.source.ts
@@ -1,0 +1,4 @@
+export enum EsdtDataSource {
+  gateway = 'gateway',
+  elastic = 'elastic'
+}
