+++ src/endpoints/nfts/entities/nft.query.options.ts
@@ -3,5 +3,4 @@ export class NftQueryOptions {
   withSupply?: boolean = false;
   withTimestamp?: boolean = false;
   withMetadata?: boolean = false;
-  source?: string = 'gateway';
 }
