+++ src/endpoints/nfts/nft.service.ts
@@ -22,7 +22,6 @@ import { QueryType } from "src/common/elastic/entities/query.type";
 import { QueryOperator } from "src/common/elastic/entities/query.operator";
 import { CachingService } from "src/common/caching/caching.service";
 import { Constants } from "src/utils/constants";
-import { GatewayComponentRequest } from "src/common/gateway/entities/gateway.component.request";
 import { PluginService } from "src/common/plugins/plugin.service";
 import { NftMetadataService } from "src/queue.worker/nft.worker/queue/job-services/metadata/nft.metadata.service";
 import { NftMediaService } from "src/queue.worker/nft.worker/queue/job-services/media/nft.media.service";
@@ -32,6 +31,8 @@ import { CacheInfo } from "src/common/caching/entities/cache.info";
 import { PersistenceInterface } from "src/common/persistence/persistence.interface";
 import { RecordUtils } from "src/utils/record.utils";
 import { EsdtSupply } from "../esdt/entities/esdt.supply";
+import { EsdtDataSource } from "../esdt/entities/esdt.data.source";
+import { NftAccountService } from "./nft.account.service";
 
 @Injectable()
 export class NftService {
@@ -40,10 +41,9 @@ export class NftService {
   private readonly DEFAULT_MEDIA: NftMedia[];
 
   constructor(
-    private readonly gatewayService: GatewayService,
     private readonly apiConfigService: ApiConfigService,
     private readonly elasticService: ElasticService,
-    private readonly nftExtendedAttributesService: NftExtendedAttributesService,
+    private readonly nftAccountService: NftAccountService,
     private readonly esdtService: EsdtService,
     private readonly tokenAssetService: TokenAssetService,
     private readonly cachingService: CachingService,
@@ -389,30 +389,13 @@ export class NftService {
     return await this.elasticService.getCount('tokens', elasticQuery);
   }
 
-  async getNftsForAddress(address: string, queryPagination: QueryPagination, filter: NftFilter, queryOptions?: NftQueryOptions): Promise<NftAccount[]> {
+  async getNftsForAddress(address: string, queryPagination: QueryPagination, filter: NftFilter, queryOptions?: NftQueryOptions, source?: EsdtDataSource): Promise<NftAccount[]> {
     const { from, size } = queryPagination;
 
     let nfts;
-    nfts = await this.getNftsForAddressInternal(address, filter, queryOptions?.source);
+    nfts = await this.nftAccountService.getNftsForAddressInternal(address, filter, source);
 
     nfts = nfts.slice(from, from + size);
-
-    const identifiers = nfts.map(x => x.identifier);
-    const queries = identifiers.map(identifier =>
-      QueryType.Match('identifier', identifier, QueryOperator.AND)
-    );
-
-    const elasticQuery = ElasticQuery.create()
-      .withCondition(QueryConditionOptions.should, queries);
-
-    const elasticNfts = await this.elasticService.getList('tokens', 'identifier', elasticQuery);
-
-    for (const nft of nfts) {
-      const elasticNft = elasticNfts.find((x: any) => x.identifier === nft.identifier);
-      if (elasticNft) {
-        nft.timestamp = elasticNft.timestamp;
-      }
-    }
 
     for (const nft of nfts) {
       await this.applyAssetsAndTicker(nft);
@@ -432,158 +415,16 @@ export class NftService {
   }
 
   async getNftCountForAddress(address: string, filter: NftFilter): Promise<number> {
-    const nfts = await this.getNftsForAddressInternal(address, filter);
+    const nfts = await this.nftAccountService.getNftsForAddressInternal(address, filter);
 
     return nfts.length;
-  }
-
-  private async filterNfts(filter: NftFilter, nfts: NftAccount[]): Promise<NftAccount[]> {
-    if (filter.search) {
-      const searchLower = filter.search.toLowerCase();
-
-      nfts = nfts.filter(x => x.name.toLowerCase().includes(searchLower) || x.identifier.toLowerCase().includes(searchLower));
-    }
-
-    if (filter.type) {
-      const types = filter.type;
-
-      nfts = nfts.filter(x => types.includes(x.type));
-    }
-
-    if (filter.collection) {
-      nfts = nfts.filter(x => x.collection === filter.collection);
-    }
-
-    if (filter.name) {
-      const searchedNameLower = filter.name.toLowerCase();
-
-      nfts = nfts.filter(x => x.name.toLowerCase().includes(searchedNameLower));
-    }
-
-    if (filter.collections) {
-      const collectionsArray = filter.collections;
-      nfts = nfts.filter(x => collectionsArray.includes(x.collection));
-    }
-
-    if (filter.tags) {
-      const tagsArray = filter.tags;
-      nfts = nfts.filter(nft => tagsArray.filter(tag => nft.tags.includes(tag)).length === tagsArray.length);
-    }
-
-    if (filter.creator) {
-      nfts = nfts.filter(x => x.creator === filter.creator);
-    }
-
-    if (filter.hasUris === true) {
-      nfts = nfts.filter(x => x.uris && x.uris.length > 0);
-    } else if (filter.hasUris === false) {
-      nfts = nfts.filter(x => x.uris && x.uris.length === 0);
-    }
-
-    return nfts;
-  }
-
-  async getGatewayNfts(address: string, filter: NftFilter, source?: string): Promise<GatewayNft[]> {
-    if (filter.identifiers !== undefined) {
-      const identifiers = filter.identifiers;
-      if (identifiers.length === 1) {
-        const identifier = identifiers[0];
-        const collectionIdentifier = identifier.split('-').slice(0, 2).join('-');
-        const nonce = parseInt(identifier.split('-')[2], 16);
-
-        const { tokenData: gatewayNft } = await this.gatewayService.get(`address/${address}/nft/${collectionIdentifier}/nonce/${nonce}`, GatewayComponentRequest.addressNftByNonce);
-
-        // normalizing tokenIdentifier since it doesn't contain the nonce in this particular scenario
-        gatewayNft.tokenIdentifier = identifier;
-
-        if (gatewayNft.balance === '0') {
-          return [];
-        }
-
-        return [gatewayNft];
-      }
-
-      if (identifiers.length > 1) {
-        const esdts = await this.esdtService.getAllEsdtsForAddress(address, source);
-        return Object.values(esdts).map(x => x as any).filter(x => identifiers.includes(x.tokenIdentifier));
-      }
-    }
-
-    const esdts = await this.esdtService.getAllEsdtsForAddress(address);
-    return Object.values(esdts).map(x => x as any).filter(x => x.tokenIdentifier.split('-').length === 3);
-  }
-
-  async getNftsForAddressInternal(address: string, filter: NftFilter, source?: string): Promise<NftAccount[]> {
-    const gatewayNfts = await this.getGatewayNfts(address, filter, source);
-
-    gatewayNfts.sort((a: GatewayNft, b: GatewayNft) => a.tokenIdentifier.localeCompare(b.tokenIdentifier, 'en', { sensitivity: 'base' }));
-
-    let nfts: NftAccount[] = [];
-
-    for (const gatewayNft of gatewayNfts) {
-      const nft = new NftAccount();
-      nft.identifier = gatewayNft.tokenIdentifier;
-      nft.collection = gatewayNft.tokenIdentifier.split('-').slice(0, 2).join('-');
-      nft.nonce = gatewayNft.nonce;
-      nft.creator = gatewayNft.creator;
-      nft.royalties = Number(gatewayNft.royalties) / 100; // 10.000 => 100%
-      nft.uris = gatewayNft.uris ? gatewayNft.uris.filter((x: any) => x) : [];
-      nft.name = gatewayNft.name;
-
-      // @ts-ignore
-      delete nft.timestamp;
-
-      if (nft.uris && nft.uris.length > 0) {
-        try {
-          nft.url = TokenUtils.computeNftUri(BinaryUtils.base64Decode(nft.uris[0]), this.NFT_THUMBNAIL_PREFIX);
-        } catch (error) {
-          this.logger.error(error);
-        }
-      }
-
-      nft.isWhitelistedStorage = nft.url.startsWith(this.NFT_THUMBNAIL_PREFIX);
-
-      nft.attributes = gatewayNft.attributes;
-
-      if (gatewayNft.attributes) {
-        nft.tags = this.nftExtendedAttributesService.getTags(gatewayNft.attributes);
-      }
-
-      const collectionDetails = await this.esdtService.getEsdtTokenProperties(nft.collection);
-      if (collectionDetails) {
-        // @ts-ignore
-        nft.type = collectionDetails.type;
-
-        if (nft.type === NftType.MetaESDT) {
-          nft.decimals = collectionDetails.decimals;
-          // @ts-ignore
-          delete nft.royalties;
-          // @ts-ignore
-          delete nft.uris;
-        }
-
-        if (!nft.name) {
-          nft.name = collectionDetails.name;
-        }
-      }
-
-      if ([NftType.SemiFungibleESDT, NftType.MetaESDT].includes(nft.type)) {
-        nft.balance = gatewayNft.balance;
-      }
-
-      nfts.push(nft);
-    }
-
-    nfts = await this.filterNfts(filter, nfts);
-
-    return nfts;
   }
 
   async getNftForAddress(address: string, identifier: string): Promise<NftAccount | undefined> {
     const filter = new NftFilter();
     filter.identifiers = [identifier];
 
-    const nfts = await this.getNftsForAddressInternal(address, filter);
+    const nfts = await this.nftAccountService.getNftsForAddressInternal(address, filter);
     if (nfts.length === 0) {
       return undefined;
     }
